add periphery /terminal/execute/container api
